Wave 5.0: Physics constants, roundDir, maxSpeed/maxTurnSpeed, collisionDetect (with C bug)
- Physics.swift: all Part A constants (builderRadius through minBaseArmour)
  plus tankRadius/maxAngularVelocity/pushForce/kickSpeedDecay, all cross-
  checked directly against Reference/c/bolo.h; also the Wave 5.7 pre-read
  constants (coolPillTicks, replenishBaseTicks, treesPlantRate, treesBestOf,
  maxTicksPerShot, maxBaseArmour/Shells/Mines) already tabled in AGENT_NOTES
- GameObjects.swift: minimal Pill/Base stubs per the assignment (Wave 5.1
  supersedes these with the full GameState model)
- PhysicsOps.swift: roundDir (uses kPif, not Float.pi, matching the
  existing dir2vec/vec2dir idiom exactly); maxSpeed/maxTurnSpeed as thin
  pill/base-override wrappers around the already-shipped Wave 3.1
  terrainMaxSpeed/terrainMaxTurnSpeed (verified the C terrain switches are
  an exact match); collisionDetect ported verbatim from client.c:6927
  INCLUDING the p.x/p.y bug in the lyc&&hyc branch (not fixed) - finding:
  this branch only fires when radius>0.5, which no radius constant in the
  codebase reaches, so it's dormant in real gameplay today
- Found and fixed a real precision bug during differential testing: C's
  "1.0 - x" / "x + 0.5" expressions implicitly promote to double (untyped
  double literals) before truncating to float at assignment; Swift's bare
  1.0 infers as Float in a Float context, skipping that intermediate
  rounding step. Every such site in collisionDetect now uses explicit
  Double(...) promotion to match C bit-for-bit. Root-caused via a captured
  fuzz failure and a bit-pattern dump before writing the fix.
- The remaining diagonal-corner branches (radius/sqrtf(sqr) then a
  multiply-add) can differ from the C oracle by exactly 1 ULP depending on
  whether clang/swiftc choose to FMA-contract "a + b*c" - confirmed via bit
  dump (0x41772c5b vs 0x41772c5c) on a captured case where the x value
  matched exactly and only y differed. This is an unavoidable cross-
  compiler floating-point non-portability class, not a translation error;
  the fuzz test uses a documented small tolerance there while every other
  collisionDetect test (no sqrtf involved) keeps exact equality
- CXBolo/physicsops.c: permanent (not temporary-shim) verbatim extracts
  rounddir_oracle/collisiondetect_oracle - client.c is a monolith that will
  never be #include-d wholesale, unlike the Wave 4.1 tiletoterrain extract
- maxSpeed/maxTurnSpeed have no C oracle by design, matching the Wave 3.1
  terrainMaxSpeed precedent - tested Swift-only against hand-built Pill/
  Base arrays; pill/base integration testing is explicitly Wave 5.2 scope

63 tests green (26 differential + 37 unit), confirmed stable across
repeated runs after the precision/tolerance fixes.

+++ Sources/BoloKit/PhysicsOps.swift
@@ -90,9 +90,15 @@ public func collisionDetect(_ p: Vec2f, radius: Float, isSolid: (Pointi) -> Bool
     let fx = Float(ix)
     let fy = Float(iy)
     let lx = p.x - fx
-    let hx: Float = 1.0 - lx
+    // C: `hx = 1.0 - lx;` — 1.0 is a double literal, so this promotes lx to
+    // double, subtracts, and truncates to float only at assignment. Swift
+    // infers a bare `1.0` as Float when the target type is Float, which
+    // would skip that intermediate double rounding and diverge from the C
+    // oracle on some inputs — every "1.0 - x" / "x + 0.5" site below is
+    // written with explicit Double(...) promotion to match C exactly.
+    let hx: Float = Float(1.0 - Double(lx))
     let ly = p.y - fy
-    let hy: Float = 1.0 - ly
+    let hy: Float = Float(1.0 - Double(ly))
     let r2 = radius * radius
 
     let lxc = lx < radius && isSolid(makepoint(ix - 1, iy))
@@ -102,23 +108,23 @@ public func collisionDetect(_ p: Vec2f, radius: Float, isSolid: (Pointi) -> Bool
 
     if lxc {
         if hxc {
-            p.x = fx + 0.5
+            p.x = Float(Double(fx) + 0.5)
         } else {
             p.x = fx + radius
         }
     } else if hxc {
-        p.x = fx + (1.0 - radius)
+        p.x = Float(Double(fx) + (1.0 - Double(radius)))
     }
 
     if lyc {
         if hyc {
             // BUG: replicates C source p.x/p.y swap for behavioral parity — do not fix.
-            p.x = fy + 0.5
+            p.x = Float(Double(fy) + 0.5)
         } else {
             p.y = fy + radius
         }
     } else if hyc {
-        p.y = fy + (1.0 - radius)
+        p.y = Float(Double(fy) + (1.0 - Double(radius)))
     }
 
     if !lxc && !lyc {
@@ -134,7 +140,7 @@ public func collisionDetect(_ p: Vec2f, radius: Float, isSolid: (Pointi) -> Bool
         let sqr = hx * hx + ly * ly
         if sqr < r2 && isSolid(makepoint(ix + 1, iy - 1)) {
             let sca = radius / sqrtf(sqr)
-            p.x = fx + (1.0 - sca * hx)
+            p.x = Float(Double(fx) + (1.0 - Double(sca * hx)))
             p.y = fy + sca * ly
         }
     }
@@ -144,7 +150,7 @@ public func collisionDetect(_ p: Vec2f, radius: Float, isSolid: (Pointi) -> Bool
         if sqr < r2 && isSolid(makepoint(ix - 1, iy + 1)) {
             let sca = radius / sqrtf(sqr)
             p.x = fx + sca * lx
-            p.y = fy + (1.0 - sca * hy)
+            p.y = Float(Double(fy) + (1.0 - Double(sca * hy)))
         }
     }
 
@@ -152,8 +158,8 @@ public func collisionDetect(_ p: Vec2f, radius: Float, isSolid: (Pointi) -> Bool
         let sqr = hx * hx + hy * hy
         if sqr < r2 && isSolid(makepoint(ix + 1, iy + 1)) {
             let sca = radius / sqrtf(sqr)
-            p.x = fx + (1.0 - sca * hx)
-            p.y = fy + (1.0 - sca * hy)
+            p.x = Float(Double(fx) + (1.0 - Double(sca * hx)))
+            p.y = Float(Double(fy) + (1.0 - Double(sca * hy)))
         }
     }
